requirements/development.txt
# Development requirements
-r base.txt

piprot
django-debug-toolbar==1.4
django-extensions==1.6.7
ipython
#Markdown==2.6.6
pip-tools==1.6.5
Werkzeug==0.11.9
certifi==2016.2.28
ecdsa==0.13
Fabric==1.11.1
paramiko==2.0.0
pycrypto==2.6.1
formic==0.9beta8
# Needed during build/deployment
Jinja2==2.8
coloredlogs==5.0
django-nose==1.4.3
nose==1.3.7
factory_boy==2.7.0
sure==1.2.24
mock==2.0.0
nose-parameterized==0.5.0
yanc
